Create Drawio structure
