+++ .gitignore
@@ -7,4 +7,8 @@ data/*.json
 data/*.pdf
 data/*.xlsx
 images/
+
+test_code.py
 .pytest_cache
+*.drawio
+
