+++ image_processing/__init__.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 from image_processing.enumerates import *
 from image_processing.imageDesign import ImagesDesign
+from image_processing.DrawioProcessing import DrawioImageDesign
 
 
 __all__ = ["LayoutMode",
 		   "LabelMode",
 		   "SignaturePosition",
-		   "ImagesDesign",]
+		   "ImagesDesign",
+		   "DrawioImageDesign"]
